Extracted LoadableEncryptorInterface into its own file.
